Defer stabilization of the `uv python install --default` behavior (#14681)
I'm a little worried about the semantics of this flag and the "default
install". We can make it more accessible while stabilizing the rest of
the behaviors, while deferring final stabilization of this to a future
release. I left some TODOs in the code covering open questions for the
feature.

This is stacked on #14626, which initially stabilized this feature.