Add a reusable path-or-URL parser
Reviewing #14687, I noticed that we had implemented a `Url::from_url_or_path`-like function, but it wasn't reusable. This change `Verbatim::from_url_or_path` so we can use it in other places too.

The PEP 508 parser is an odd place for this, but that's where `VerbatimUrl` and `Scheme` are already living.