Fix Error on tool removal and remove self-replace
self-replace is no longer needed here, since we invoke the
remove_virtualenv function from uv-virtualenv.
This also caused a problem that the uv_tool::Error::Io is no longer
raised, but the uv_tool::Error::VirtualEnvError, when e.g. `black`
is not present as a tool in the test case.